Scrollbar colors adapt to mode Menu button hover color fixed Menu outlining Buttons work correctly with tab Setting buttons dont force to new page Text selection color adapt to mode Fixed missed word Pressing logo brings you to homepage Added Traditional Chinese

+++ RazorComponents/RazorComponents/Helpers/Color.cs
@@ -52,7 +52,7 @@ public record struct Color
 		this.Value = rgbColor;
 	}
 
-	public Color ChangeBrightness(sbyte delta = 40)
+	public Color ChangeBrightness(int delta = 40)
 	{
 		var r = ChangeBrightnessOfPrimaryColor(this.Value.R);
 		var g = ChangeBrightnessOfPrimaryColor(this.Value.G);
